centraliza exibição dos dados da organização na Sidebar
- Remove duplicação da exibição dos dados da organização das telas de importação e verificação
- Centraliza a exibição do card da organização na Sidebar
- Melhora a consistência visual da aplicação
- Remove título desnecessário da tela de importação
