+++ components/Sidebar.tsx
@@ -12,9 +12,26 @@ import {
   SidebarMenuButton
 } from '@/components/ui/sidebar'
 import { Home, Users } from 'lucide-react'
+import { useEffect, useState } from 'react'
+import { getOrganizationDetails } from '@/lib/api'
+import type { OrganizationDetails } from '@/lib/types'
 
 export function Sidebar() {
   const pathname = usePathname()
+  const [organization, setOrganization] = useState<OrganizationDetails | null>(null)
+
+  useEffect(() => {
+    async function loadOrganization() {
+      try {
+        const orgDetails = await getOrganizationDetails(process.env.NEXT_PUBLIC_UTALK_ORGANIZATION_ID!)
+        setOrganization(orgDetails)
+      } catch (error) {
+        console.error('Erro ao carregar organização:', error)
+      }
+    }
+    
+    loadOrganization()
+  }, [])
 
   return (
     <ShadcnSidebar>
@@ -48,6 +65,29 @@ export function Sidebar() {
             </SidebarMenuButton>
           </SidebarMenuItem>
         </SidebarMenu>
+
+        {/* Card da Organização */}
+        {organization && (
+          <div className="mx-2 mt-4">
+            <div className="bg-white rounded-lg border border-gray-200 shadow-sm p-4">
+              <div className="flex items-center gap-3">
+                {organization.iconUrl && (
+                  <img 
+                    src={organization.iconUrl} 
+                    alt={organization.name}
+                    className="w-8 h-8 rounded-full object-cover"
+                  />
+                )}
+                <div>
+                  <span className="font-medium">{organization.name}</span>
+                  <span className="text-sm text-gray-500 block">
+                    ID: {organization.id}
+                  </span>
+                </div>
+              </div>
+            </div>
+          </div>
+        )}
       </SidebarContent>
     </ShadcnSidebar>
   )
